backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
werkzeug>=3.0.0
PyJWT>=2.8.0
Pillow>=10.3.0
qrcode[pil]>=7.4.2
pyzbar>=0.1.9
pyodbc>=5.1.0
# SQLAlchemy MSSQL adapter (Azure SQL)
# Install the appropriate ODBC driver for your OS:
#   macOS:   brew install msodbcsql17
#   Ubuntu:  apt-get install msodbcsql17
# pyodbc is listed above; SQLAlchemy uses it via mssql+pyodbc URI
